+++ chart/tests/urls.py
@@ -1,0 +1,10 @@
+from django.conf.urls import patterns, include
+
+
+urlpatterns = patterns(
+    '',
+    (r'^jmbo/', include('jmbo.urls')),
+    (r'^comments/', include('django.contrib.comments.urls')),
+    (r'^music/', include('music.urls')),
+    (r'^chart/', include('chart.urls')),
+)