+++ chart/urls.py
@@ -1,11 +1,13 @@
-from django.conf.urls.defaults import patterns, url
+from django.conf.urls import patterns, url
+
+from jmbo.views import ObjectDetail
 
 
 urlpatterns = patterns(
     '',
     url(
-        r'^(?P<slug>[\w-]+)/$', 
-        'jmbo.views.object_detail',
+        r'^(?P<slug>[\w-]+)/$',
+        ObjectDetail.as_view(),
         name='chart_object_detail'
     ),
 )