+++ test_settings.py
@@ -1,41 +1,74 @@
-import os
+# We're still on Django 1.4 and use django-setuptest. Use this as a starting
+# point for your test settings. Typically copy this file as test_settings.py
+# and replace myapp with your app name.
+
+from os.path import expanduser
 
 DEBUG = True
-TEMPLATE_DEBUG = DEBUG
 
 DATABASES = {
     'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
+        'ENGINE': 'django.db.backends.postgresql_psycopg2',
+        'NAME': 'jmbo',
+        'USER': 'postgres',
+        'PASSWORD': '',
+        'HOST': '',
+        'PORT': '',
     }
 }
 
-ROOT_URLCONF = 'chart.urls'
-
 INSTALLED_APPS = (
-    # jmbo required apps
-    'category',
-    'photologue',
-    'secretballot',
-
-    # elective apps
-    'publisher',
-    'jmbo',
-    'category',
-    'music',
     'chart',
-
-    # django required apps
-    'django.contrib.auth',
+    'music',
+    'jmbo',
+    'photologue',
+    'category',
+    'likes',
+    'secretballot',
+    'publisher',
+    'preferences',
     'django.contrib.admin',
+    'django.contrib.auth',
+    'django.contrib.comments',
     'django.contrib.contenttypes',
+    'django.contrib.sites',
+    'south',
 )
+
+ROOT_URLCONF = 'chart.tests.urls'
 
 MIDDLEWARE_CLASSES = (
+    'django.middleware.common.CommonMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
+    'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
+    'django.contrib.messages.middleware.MessageMiddleware',
+    'likes.middleware.SecretBallotUserIpUseragentMiddleware',
 )
 
+TEMPLATE_CONTEXT_PROCESSORS = (
+    'django.contrib.auth.context_processors.auth',
+    'django.core.context_processors.debug',
+    'django.core.context_processors.i18n',
+    'django.core.context_processors.media',
+    'django.core.context_processors.static',
+    'django.core.context_processors.tz',
+    'django.core.context_processors.request',
+    'django.contrib.messages.context_processors.messages',
+)
 
-CKEDITOR_MEDIA_PREFIX = '/media/ckeditor/'
+USE_TZ = True
 
 SITE_ID = 1
+
+STATIC_URL = '/static/'
+STATIC_ROOT = 'static/'
+
+CKEDITOR_UPLOAD_PATH = expanduser('~')
+
+SOUTH_TESTS_MIGRATE = False
+
+# Disable celery
+CELERY_ALWAYS_EAGER = True
+BROKER_BACKEND = 'memory'
+