+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+2.0.0a1
+-------
+#. PEP8 cleanup.
+#. Refactor tests.
+#. Django 1.6 compatibility.
+#. Up minimum jmbo to 2.0.0.
+
 0.1
 ---
 #. Introduce dependency on `jmbo-foundry`.